Added a way to skip whitespace
Added skip_ws parameter to the constructor.
The recognize method now removes whitespace from the front of the buffer
just before it starts matching tokens.  If it matches it will restart
the loop as to reset the loop as if nothing happened.

This is a better approach than removing whitespace indiscrimnately from
the input_buffer, because whitespace can be between two similar tokens,
where removing the whitespace will paste them together.